requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.0
bcrypt==4.1.1
aiofiles==23.2.1
pillow==10.1.0

# ML Dependencies
tensorflow==2.18.0
numpy==1.26.4
scikit-learn==1.5.2
matplotlib==3.9.2
opencv-python==4.10.0.84
pandas==2.2.3

# Email and Notifications
emails==0.6
jinja2==3.1.2
twilio==8.11.0
requests==2.31.0
